+++ pandas/tests/frame/constructors/test_from_dict.py
@@ -85,7 +85,7 @@ class TestFromDict:
         expected = DataFrame.from_dict(sdict, orient="index")
         tm.assert_frame_equal(result, expected.reindex(result.index))
 
-        result2 = DataFrame(data, index=np.arange(6))
+        result2 = DataFrame(data, index=np.arange(6, dtype=np.int64))
         tm.assert_frame_equal(result, result2)
 
         result = DataFrame([Series(dtype=object)])
