+++ pandas/tests/groupby/test_filters.py
@@ -232,7 +232,7 @@ def test_filter_using_len():
     actual = grouped.filter(lambda x: len(x) > 2)
     expected = DataFrame(
         {"A": np.arange(2, 6), "B": list("bbbb"), "C": np.arange(2, 6)},
-        index=np.arange(2, 6),
+        index=np.arange(2, 6, dtype=np.int64),
     )
     tm.assert_frame_equal(actual, expected)
 
@@ -244,7 +244,7 @@ def test_filter_using_len():
     s = df["B"]
     grouped = s.groupby(s)
     actual = grouped.filter(lambda x: len(x) > 2)
-    expected = Series(4 * ["b"], index=np.arange(2, 6), name="B")
+    expected = Series(4 * ["b"], index=np.arange(2, 6, dtype=np.int64), name="B")
     tm.assert_series_equal(actual, expected)
 
     actual = grouped.filter(lambda x: len(x) > 4)
